First method implemented
For the first part of the project, I created a python code that generates a LUT table of function x2 - x3 for x in range (0, 1) and the table has 256 sluts which means we only need 8-bits so we create LUT for every possible number in 8-bits and also, they are less than 1 and bigger than 0 so we use all 8-bits as the fraction. After generating the LUT table we only need to make a module that returns the related LUT entry based on 8-bit input.