linear interpolation method and test implemented
For the second part, we must implement the function using linear interpolation using 64 slut memories so first of all, we must separate 64 entries from the table and the choices must be in equal places so we collect every 4k (k = 0, 1, ...) entry. After that we need to implement the formula below inside our Verilog module. For being able to synthesize the module on FPGA we must remove the division; the best thing is to make it into a shift division (base 2 division) so that let module gets synthesized. Base on the course slides we can convert the formula to this (we know that XLSB is input (8-6=)2 LSB bits and X2 – X1 = 2N-P = 2(8-6)).